[#388] Refactor: 일기 관련 엔드포인트 수정
