+++ src/main/java/umc/GrowIT/Server/service/diaryService/DiaryQueryService.java
@@ -4,8 +4,6 @@ import umc.GrowIT.Server.web.dto.DiaryDTO.DiaryResponseDTO;
 
 public interface DiaryQueryService {
 
-    boolean hasVoiceDiaries(Long userId);
-
     DiaryResponseDTO.DiaryDateListDTO getDiaryDate(Integer year, Integer month, Long userId);
 
     DiaryResponseDTO.DiaryListDTO getDiaryList(Integer year, Integer month, Long userId);
