+++ src/main/java/umc/GrowIT/Server/service/diaryService/DiaryQueryServiceImpl.java
@@ -5,14 +5,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import umc.GrowIT.Server.apiPayload.code.status.ErrorStatus;
 import umc.GrowIT.Server.apiPayload.exception.DiaryHandler;
-import umc.GrowIT.Server.apiPayload.exception.UserHandler;
 import umc.GrowIT.Server.converter.DiaryConverter;
 import umc.GrowIT.Server.domain.Diary;
-import umc.GrowIT.Server.domain.User;
-import umc.GrowIT.Server.domain.enums.CreditSource;
-import umc.GrowIT.Server.repository.CreditHistoryRepository;
 import umc.GrowIT.Server.repository.DiaryRepository;
-import umc.GrowIT.Server.repository.UserRepository;
 import umc.GrowIT.Server.web.dto.DiaryDTO.DiaryResponseDTO;
 
 import java.util.List;
@@ -24,15 +19,6 @@ import java.util.Optional;
 public class DiaryQueryServiceImpl implements DiaryQueryService{
 
     private final DiaryRepository diaryRepository;
-    private final CreditHistoryRepository creditHistoryRepository;
-    private final UserRepository userRepository;
-
-    @Override
-    public boolean hasVoiceDiaries(Long userId) {
-        User user = userRepository.findById(userId).orElseThrow(() -> new UserHandler(ErrorStatus.USER_NOT_FOUND));
-
-        return creditHistoryRepository.existsByUserAndSource(user, CreditSource.VOICE_DIARY);
-    }
 
     @Override
     public DiaryResponseDTO.DiaryDateListDTO getDiaryDate(Integer year, Integer month, Long userId) {
