+++ src/main/java/umc/GrowIT/Server/service/userService/UserQueryService.java
@@ -11,4 +11,6 @@ public interface UserQueryService {
     UserResponseDTO.CreditHistoryResponseDTO getCreditHistory(Long userId, int year, int month, CreditTransactionType type, int page);
 
     UserResponseDTO.EmailResponseDTO getMyEmail(Long userId);
+
+    boolean hasVoiceDiaries(Long userId);
 }
