+++ src/main/java/umc/GrowIT/Server/service/userService/UserQueryServiceImpl.java
@@ -12,6 +12,7 @@ import umc.GrowIT.Server.converter.UserConverter;
 import umc.GrowIT.Server.domain.CreditHistory;
 import umc.GrowIT.Server.domain.Gro;
 import umc.GrowIT.Server.domain.User;
+import umc.GrowIT.Server.domain.enums.CreditSource;
 import umc.GrowIT.Server.repository.CreditHistoryRepository;
 import umc.GrowIT.Server.repository.GroRepository;
 import umc.GrowIT.Server.repository.UserRepository;
@@ -75,6 +76,12 @@ public class UserQueryServiceImpl implements UserQueryService {
         return UserConverter.toEmailResponseDTO(user);
     }
 
+    @Override
+    public boolean hasVoiceDiaries(Long userId) {
+        User user = userRepository.findById(userId).orElseThrow(() -> new UserHandler(ErrorStatus.USER_NOT_FOUND));
+
+        return creditHistoryRepository.existsByUserAndSource(user, CreditSource.VOICE_DIARY);
+    }
 
     /*
         이 아래부터 헬퍼메소드
