+++ src/main/java/umc/GrowIT/Server/web/controller/specification/DiarySpecification.java
@@ -19,12 +19,6 @@ import umc.GrowIT.Server.web.dto.DiaryDTO.DiaryRequestDTO;
 import umc.GrowIT.Server.web.dto.DiaryDTO.DiaryResponseDTO;
 
 public interface DiarySpecification {
-    @GetMapping("/has-voice-diary")
-    @Operation(summary = "음성 일기 작성 여부 조회 API",description = "사용자가 한번이라도 음성 일기를 작성했는지의 여부를 조회하는 API입니다.")
-    @ApiResponses({
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON_200", description = "⭕ SUCCESS"),
-    })
-    ApiResponse<Boolean> hasVoiceDiary();
 
     @GetMapping("/dates")
     @Operation(summary = "일기 작성 날짜 조회 API",description = "특정 사용자의 일기 메인 화면에서 사용할 월별 일기 기록한 날짜를 보여주기 위한 API입니다. Query String으로 year와 month를 주세요." +
@@ -39,7 +33,7 @@ public interface DiarySpecification {
             @Parameter(description = "일기 작성 월",
                     example = "9")@RequestParam Integer month);
 
-    @GetMapping("/")
+    @GetMapping("")
     @Operation(summary = "일기 모아보기 API",description = "특정 사용자가 작성한 일기를 모아보는 API입니다. query string으로 year와 month를 넘겨주면 해당 월에 작성한 일기들의 리스트, 작성한 일기 수를 보내줍니다.")
     @ApiResponses({
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON_200", description = "⭕ SUCCESS"),
